Refactoring - organized routes and separated controller logic
