+++ app/views/AssignTeams.ejs
@@ -54,8 +54,41 @@
       </button>
 
       <script>
-        const results_body = document.querySelector("#results");
+        let html = "";
+        const students =  <%- JSON.stringify(result[0]) %>;
 
+        if (students.length > 0) {
+          students.forEach((student) => {
+            if (typeof student.Team === "object") {
+              var team = "-";
+            } else {
+              var team = student.Team;
+            }
+            html += `
+              <tr>
+                <td>`+student.ID+`</td>
+                <td style="text-align:left">`+student.Username+`</td>
+                <td contenteditable
+                    onclick="this.innerHTML=''"
+                    onblur="update_data(this,
+                                        '`+student.Username+`',
+                                        `+student.ID+`,
+                                        '`+team+`')">`
+                    +team+`</td>
+              </tr>`;
+          });
+        } else {
+            html += `
+              <tr>
+                <td colspan="3"><br>
+                  <h2>No students in the system. Please upload the course roster first.</h2>
+                </td>
+              </tr?>`;
+        }
+
+        document.querySelector("#results").innerHTML = html;
+
+        // Auto-assign teams based on teacher's input of number of students per team
         document.querySelector("form").onsubmit = (event) => {
           event.preventDefault();
 
@@ -73,6 +106,7 @@
           });
         };
 
+        // Prevents return key from being registered as part of a team name
         document
           .getElementById("results")
           .addEventListener("keydown", (evt) => {
@@ -82,67 +116,7 @@
             }
           });
 
-        load_data();
-
-        function load_data() {
-          const request = new XMLHttpRequest();
-
-          request.open(`GET`, `/get-students`);
-
-          let html = "";
-
-          request.onreadystatechange = () => {
-            if (
-              request.readyState === XMLHttpRequest.DONE &&
-              request.status === 200
-            ) {
-              const results = JSON.parse(request.responseText);
-              if (results.length > 0) {
-                results.forEach((result) => {
-                  if (typeof result.Team === "object") {
-                    var team = "-";
-                  } else {
-                    var team = result.Team;
-                  }
-                  html +=
-                    `<tr>
-                                    <td>` +
-                    result.ID +
-                    `</td>
-                                    <td style="text-align:left">` +
-                    result.Username +
-                    `</td>
-                                    <td contenteditable
-                                        onclick="this.innerHTML=''"
-                                        onblur="update_data(this,
-                                                            '` +
-                    result.Username +
-                    `',
-                                                            ` +
-                    result.ID +
-                    `,
-                                                            '` +
-                    team +
-                    `')">` +
-                    team +
-                    `</td>
-                            </tr>`;
-                });
-              } else {
-                html += `<tr><td colspan="3">
-                                <br>
-                                <h2>No students in the system. Please upload the course roster first.</h2>
-                            </td>
-                        </tr?>`;
-              }
-
-              results_body.innerHTML = html;
-            }
-          };
-
-          request.send();
-        }
-
+        // Assigns an individual student's team based on teacher's input
         function update_data(element, name, id, currentTeam) {
           var input = element.innerHTML;
           if (input.length > 0 && input != currentTeam) {
