+++ app/views/Login.ejs
@@ -14,7 +14,7 @@
 
   <body>
     <div class="wrapper">
-      <form action="/LogUser" method="POST" class="LoginForm">
+      <form action="/login" method="POST" class="LoginForm">
         <h2>Login Page</h2>
 
         <div class="input-box">
@@ -46,4 +46,5 @@
         <button type="submit" value="Submit" class="btn">Login</button>
       </form>
     </div>
-  </body></HTML>
+  </body></HTML
+>
