+++ app/views/MainPage.ejs
@@ -32,7 +32,7 @@
       <% if (Username && Role==="Student" ) { %>
       <div class="button-container">
         <div class="wrapper">
-          <button class="btn" onclick="window.location.href='/TeamVis'">
+          <button class="btn" onclick="window.location.href='/teammates'">
             View Teammates
           </button>
           <button class="btn" onclick="window.location.href='/logout'">
@@ -49,7 +49,7 @@
           <button class="btn" onclick="window.location.href='/assign-teams'">
             Assign Teams
           </button>
-          <button class="btn" onclick="window.location.href='/AllTeamVis'">
+          <button class="btn" onclick="window.location.href='/teams'">
             View Teams
           </button>
           <button class="btn" onclick="window.location.href='/logout'">
